requirements.txt
# Minimal, up-to-date requirements for this Flask app
# Only include top-level packages actually used by the project.
Flask>=2.2.3,<3.0
python-dotenv>=0.21.0

keras>=3.0
tensorflow>=2.16.1
google-cloud-storage>=2.7.0


# Testing
pytest>=7.2
numpy>=1.24
opencv-python

# Install from a virtual environment:
# python -m venv .venv; .\.venv\Scripts\Activate.ps1; pip install -r requirements.txt